Move definitions related to Streamlit version compatibility features to _compat.py
